day 12, part 1: lcm(cycle_time) will find total cycle time
Each axis repeats on its own cycle, which is way shorter than
the hundreds of trillions of possible solutions generated by my
input.

The future of a particular axis is deterministic, so there can be
no sub-cycles.

I could save a bit of time by calculating the axes column-wise instead
of re-simulating the entire universe three times, but it already runs
in just a few seconds, so I'm not going to fix it.